+++ src/main/kotlin/de/tu_chemnitz/fft/data/FFT.kt
@@ -19,7 +19,7 @@ value class Bins(private val bins: Int) {
 data class FFTData(
     val bins: Bins,
     val sampleSize: Int,
-    val samplingRate: Double,
+    val samplingRate: Int,
     val output: List<Complex>
 ) {
     private fun abs(n: Complex): Complex = sqrt(n.re * n.re + n.im * n.im)
@@ -27,7 +27,6 @@ data class FFTData(
     fun magnitude() = output.map { abs(it).re / sampleSize }
 
     // index = Frequency * Number of FFT Points / Sampling Frequency
-    //
     fun binIndexOf(frequency: Double): Int = (frequency * sampleSize / samplingRate).roundToInt()
 }
